Address review: keep data source read in datasource.go, guard nullable time/int
- Move Read() and state helpers back into datasource.go (removing read.go),
  consistent with the established convention for framework data sources.
- Use convert.TimeToType for timestamp fields instead of calling .String()
  directly (RFC3339 formatting; nil-safe) in networkpoolserver, monitoringcheck
  and monitoringgroup.
- Guard the pointer returned by NullableInt32.Get() before dereferencing it in
  networkpoolserver (service_port).
- Drop redundant 'computed_optional_required: computed' overrides from data
  source config.yaml files; the generator already defaults these to computed
  (schemas unchanged).
