+++ specs/datasources/backup_instance/config.yaml
@@ -27,21 +27,16 @@ backup_instance:
       computed_optional_required: computed_optional
     backup_provider:
       description: The backup provider object.
-      computed_optional_required: computed
     backup_repository:
       description: The backup repository object.
-      computed_optional_required: computed
     backup_type:
       description: The backup type object.
-      computed_optional_required: computed
     container_id:
       description: The ID of the container (workload) being backed up
     cron_expression:
       description: The cron expression for the backup schedule.
-      computed_optional_required: computed
     date_created:
       description: The date the backup was created.
-      computed_optional_required: computed
     enabled:
       description: Whether the backup is enabled
     instance_id:
@@ -50,39 +45,27 @@ backup_instance:
       description: The ID of the Backup Job associated with this backup
     last_updated:
       description: The date the backup was last updated.
-      computed_optional_required: computed
     location:
       description: The backup storage location.
-      computed_optional_required: computed
     location_type:
       description: The backup storage location type.
-      computed_optional_required: computed
     retention_count:
       description: The number of backups to retain.
-      computed_optional_required: computed
     schedule:
       description: The backup schedule object.
-      computed_optional_required: computed
     storage_provider_id:
       description: The ID of the storage provider (Bucket or File Share) used for this backup
     target_all:
       description: Whether to target all databases.
-      computed_optional_required: computed
     target_host:
       description: The target database host.
-      computed_optional_required: computed
     target_name:
       description: The target database name.
-      computed_optional_required: computed
     target_path:
       description: The file or directory path being backed up.
-      computed_optional_required: computed
     target_port:
       description: The target database port.
-      computed_optional_required: computed
     target_username:
       description: The target database username.
-      computed_optional_required: computed
     volume_path:
       description: The LVM volume path.
-      computed_optional_required: computed
