+++ specs/datasources/certificate/config.yaml
@@ -17,16 +17,12 @@ certificate:
       conflicts_with: ["id"]
     account_id:
       description: The account ID associated with the certificate.
-      computed_optional_required: computed
     category:
       description: The category of the certificate.
-      computed_optional_required: computed
     cert_type:
       description: The certificate type (e.g. server, authority).
-      computed_optional_required: computed
     common_name:
       description: The common name (CN) of the certificate.
-      computed_optional_required: computed
     description:
       description: A description of the certificate.
     domain_name:
@@ -35,19 +31,13 @@ certificate:
       description: Whether the certificate is enabled.
     generated:
       description: Whether the certificate was generated by Morpheus.
-      computed_optional_required: computed
     integration_id:
       description: The integration ID associated with the certificate.
-      computed_optional_required: computed
     self_signed:
       description: Whether the certificate is self-signed.
-      computed_optional_required: computed
     type.id:
       description: The certificate type object ID.
-      computed_optional_required: computed
     type.code:
       description: The certificate type object code.
-      computed_optional_required: computed
     wildcard:
       description: Whether the certificate is a wildcard certificate.
-      computed_optional_required: computed
