+++ specs/datasources/vdi_app/config.yaml
@@ -12,17 +12,13 @@ vdi_app:
       conflicts_with: ["id"]
     date_created:
       description: The date the VDI app was created.
-      computed_optional_required: computed
     description:
       description: The description of the VDI app.
     icon_path:
       description: The icon path for the VDI app.
-      computed_optional_required: computed
     last_updated:
       description: The date the VDI app was last updated.
-      computed_optional_required: computed
     launch_prefix:
       description: The RDS app name prefix.
     logo:
       description: The logo path for the VDI app.
-      computed_optional_required: computed
